refs #12193 Changed type for semester number field to integer

+++ src/main/java/ru/nsu/ostest/adapter/in/rest/model/laboratory/LaboratorySearchRequestDto.java
@@ -1,4 +1,4 @@
 package ru.nsu.ostest.adapter.in.rest.model.laboratory;
 
-public record LaboratorySearchRequestDto(Boolean isHidden, Byte semesterNumber) {
+public record LaboratorySearchRequestDto(Boolean isHidden, Integer semesterNumber) {
 }
